structure change for pypi packaging, turn off setWordWrap
Move everything to a MeiTingTrunk folder.
at top level, add setup.py and LICENCE.

inside MeiTingTrunk, add version.py containing the

```
__version__='current_version'
```

and imoprt that from main.py and _MainWindow.py.
in setup.py, read version.py and use exec() to get the version.

the relative import rules are very confusing. Remember to use

```
python -m MeiTingTrunk.main
```

to run the code in the future.

Tried building and uploading to pypi and seems to be working, except the
texts are displayed as wrapped in the doc table. It turns out that this
is a default change somewhere between PyQt5.6 and PyQt5.12. Therefore
in _MainFrame.py, add setWordWrap(False).

Also specifiy the pdfminer dependency to be pdfminer.six, for
compatibility with python3.